+++ client.go
@@ -131,13 +131,21 @@ func New(clientID, clientSecret, fingerprint, ipAddress string, modes ...bool) *
 		}
 	}
 
+	request := Request{
+		authKey:      "",
+		clientID:     clientID,
+		clientSecret: clientSecret,
+		fingerprint:  fingerprint,
+		ipAddress:    ipAddress,
+	}
+
 	log.info("Building new client...")
 	return &Client{
 		ClientID:     clientID,
 		ClientSecret: clientSecret,
 		Fingerprint:  fingerprint,
 		IP:           ipAddress,
-		request:      Request{"", clientID, clientSecret, fingerprint, ipAddress},
+		request:      request,
 	}
 }
 
@@ -213,6 +221,14 @@ func (c *Client) LocateATMs(queryParams ...string) (map[string]interface{}, erro
 	return c.do("GET", url, "", queryParams)
 }
 
+// VerifyAddress checks if an address if valid
+func (c *Client) VerifyAddress(data string) (map[string]interface{}, error) {
+	url := buildURL("address-verification")
+
+	log.info("Verifying address...")
+	return c.do("POST", url, data, nil)
+}
+
 // VerifyRoutingNumber checks and returns the bank details of a routing number
 func (c *Client) VerifyRoutingNumber(data string) (map[string]interface{}, error) {
 	url := buildURL("routing-number-verification")
@@ -284,20 +300,24 @@ func (c *Client) GetUsers(queryParams ...string) (map[string]interface{}, error)
 }
 
 // GetUser returns a single user
-func (c *Client) GetUser(userID string, fingerprint, ipAddress string, queryParams ...string) (*User, error) {
+func (c *Client) GetUser(userID string, queryParams ...string) (*User, error) {
 	url := buildURL(path["users"], userID)
 	res, err := c.do("GET", url, "", queryParams)
 
 	var user User
 	mapstructure.Decode(res, &user)
 	user.Response = res
-	user.request = Request{
-		"",
-		c.ClientID,
-		c.ClientSecret,
-		fingerprint,
-		ipAddress,
+
+	request := Request{
+		authKey:      "",
+		clientID:     c.ClientID,
+		clientSecret: c.ClientSecret,
+		fingerprint:  c.Fingerprint,
+		ipAddress:    c.IP,
 	}
+
+	user.request = request
+
 	log.info("Getting user...")
 	return &user, err
 }